Do not use the write_url for datastore_search
This commit remove the use of the write_url in the datastore_search
action, as the permissions are already checked against the
auth.datastore_search function (which delegates to resource_show).
Looks like the non-removal of write_url in the action method was an
accidental oversight.

Tests (including the privacy tests for TestDatastoreSearch.test_search_private_dataset)
continue to pass.